Organized files into folders
Both organized, refactored, and updated references in relevant files.
