+++ main.py
@@ -1,4 +1,4 @@
-from ui import run
+from frontend.ui import run
 
 if __name__ == '__main__':
   run()
